+++ tests/test_data.py
@@ -5,54 +5,55 @@ import fakeredis
 import hashlib
 from acumos_proto_viewer.utils import register_proto_from_url
 from acumos_proto_viewer import data
+import test_constants
 
 
 def test_msg_to_json_preserve_bytes(monkeypatch, monkeyed_requests_get, cleanuptmp, fake_msg_as_jsonwb, fake_msg, fake_msg_with_arrays_jsonwb, fake_msg_with_arrays):
     monkeypatch.setattr('requests.get', monkeyed_requests_get)
-    register_proto_from_url("http://myserver.com/fakemodelid/1.0.0/fakemodelid-1.0.0-proto")
-    register_proto_from_url("http://myserver.com/fakemodelidwitharrays/1.0.0/fakemodelidwitharrays-1.0.0-proto")
+    register_proto_from_url(test_constants.test_proto_url)
+    register_proto_from_url(test_constants.test_proto_with_arrays_url)
     msgb = fake_msg()
     # the method under test here generates a timestamp so we have to fake that
     monkeypatch.setattr('time.time', lambda: 55555555555)
     jpb = data._msg_to_json_preserve_bytes(
-        msgb, "fakemodelid_100_proto", "Data1", 1)
+        msgb, test_constants.test_proto_mid, "Data1", 1)
     assert (jpb == fake_msg_as_jsonwb())
 
     msgba = fake_msg_with_arrays()
     jpb2 = data._msg_to_json_preserve_bytes(
-        msgba, "fakemodelidwitharrays_100_proto", "ImageTagSet", 1)
+        msgba, test_constants.test_proto_with_arrays_mid, "ImageTagSet", 1)
     assert(jpb2 == fake_msg_with_arrays_jsonwb())
 
     cleanuptmp()
 
 
 def _verify_inject_test(fake_msg_as_jsonwb, fake_msg_with_arrays_jsonwb):
-    expected_h_1 = "{0}{1}{2}".format("fakemodelid_100_proto", "Data1", 'asdf0')
+    expected_h_1 = "{0}{1}{2}".format(test_constants.test_proto_mid, "Data1", 'asdf0')
     expected_index_1 = hashlib.sha224(
         expected_h_1.encode('utf-8')).hexdigest().encode()
 
-    expected_h_2 = "{0}{1}{2}".format("fakemodelidwitharrays_100_proto", "ImageTagSet", 'asdf0')
+    expected_h_2 = "{0}{1}{2}".format(test_constants.test_proto_with_arrays_mid, "ImageTagSet", 'asdf0')
     expected_index_2 = hashlib.sha224(
         expected_h_2.encode('utf-8')).hexdigest().encode()
 
     assert(data.myredis.keys() == [expected_index_1, expected_index_2])
 
-    assert(data.get_raw_data("fakemodelid_100_proto",
+    assert(data.get_raw_data(test_constants.test_proto_mid,
                              "Data1", 0, 1) == [fake_msg_as_jsonwb()])
 
-    assert(data.get_raw_data("fakemodelidwitharrays_100_proto", "ImageTagSet", 0, 1) == [fake_msg_with_arrays_jsonwb()])
+    assert(data.get_raw_data(test_constants.test_proto_with_arrays_mid, "ImageTagSet", 0, 1) == [fake_msg_with_arrays_jsonwb()])
 
     data.myredis.flushall()
 
 
 def test_inject_data_env(monkeypatch, monkeyed_requests_get, cleanuptmp, fake_msg, fake_msg_with_arrays, fake_msg_as_jsonwb, fake_msg_with_arrays_jsonwb):
     monkeypatch.setattr('requests.get', monkeyed_requests_get)
-    assert('fakemodelid_100_proto' not in data.list_known_protobufs())
-    assert('fakemodelidwitharrays_100_proto' not in data.list_known_protobufs())
-    register_proto_from_url("http://myserver.com/fakemodelidwitharrays/1.0.0/fakemodelidwitharrays-1.0.0-proto")
-    register_proto_from_url("fakemodelid/1.0.0/fakemodelid-1.0.0-proto")
-    assert('fakemodelid_100_proto' in data.list_known_protobufs())
-    assert('fakemodelidwitharrays_100_proto' in data.list_known_protobufs())
+    assert(test_constants.test_proto_mid not in data.list_known_protobufs())
+    assert(test_constants.test_proto_with_arrays_mid not in data.list_known_protobufs())
+    register_proto_from_url(test_constants.test_proto_url)
+    register_proto_from_url(test_constants.test_proto_with_arrays_url)
+    assert(test_constants.test_proto_mid in data.list_known_protobufs())
+    assert(test_constants.test_proto_with_arrays_mid in data.list_known_protobufs())
 
     # patching
     data.myredis = fakeredis.FakeStrictRedis()
@@ -63,12 +64,9 @@ def test_inject_data_env(monkeypatch, monkeyed_requests_get, cleanuptmp, fake_ms
     monkeypatch.setattr('time.time', lambda: 55555555555)
 
     msgb = fake_msg()
-    data.inject_data(
-        msgb, "fakemodelid/1.0.0/fakemodelid-1.0.0-proto", "Data1")
-
+    data.inject_data(msgb, test_constants.test_proto_url, "Data1")
     msgbwa = fake_msg_with_arrays()
-    data.inject_data(
-        msgbwa, "fakemodelidwitharrays/1.0.0/fakemodelidwitharrays-1.0.0-proto", "ImageTagSet")
+    data.inject_data(msgbwa, test_constants.test_proto_with_arrays_url, "ImageTagSet")
     _verify_inject_test(fake_msg_as_jsonwb, fake_msg_with_arrays_jsonwb)
 
     cleanuptmp()
@@ -84,17 +82,14 @@ def test_inject_data(monkeypatch, monkeyed_requests_get, cleanuptmp, fake_msg, f
         'acumos_proto_viewer.data._get_bucket', lambda: 'asdf0')
     monkeypatch.setattr('time.time', lambda: 55555555555)
 
-    assert('fakemodelid_100_proto' not in data.list_known_protobufs())
-    assert('fakemodelidwitharrays_100_proto' not in data.list_known_protobufs())
+    assert(test_constants.test_proto_mid not in data.list_known_protobufs())
+    assert(test_constants.test_proto_with_arrays_mid not in data.list_known_protobufs())
     msgb = fake_msg()
-    data.inject_data(
-        msgb, "http://myserver.com/fakemodelid/1.0.0/fakemodelid-1.0.0-proto", "Data1")
+    data.inject_data(msgb, test_constants.test_proto_url, "Data1")
     msgbwa = fake_msg_with_arrays()
-    data.inject_data(
-        msgbwa, "fakemodelidwitharrays/1.0.0/fakemodelidwitharrays-1.0.0-proto", "ImageTagSet")
-
-    assert('fakemodelid_100_proto' in data.list_known_protobufs())
-    assert('fakemodelidwitharrays_100_proto' in data.list_known_protobufs())
+    data.inject_data(msgbwa, test_constants.test_proto_with_arrays_url, "ImageTagSet")
+    assert(test_constants.test_proto_mid in data.list_known_protobufs())
+    assert(test_constants.test_proto_with_arrays_mid in data.list_known_protobufs())
     _verify_inject_test(fake_msg_as_jsonwb, fake_msg_with_arrays_jsonwb)
 
     cleanuptmp()